+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 [Commits](https://github.com/thebigmunch/google-music/compare/3.3.0...master)
 
+### Fixed
+
+* Change ``HTTPError`` to import from ``httpx`` top level.
+
 
 ## [3.3.0](https://github.com/thebigmunch/google-music/releases/tag/3.3.0) (2020-01-11)